fix warning

+++ src/test/kerntest.c
@@ -31,10 +31,10 @@ void gerepilemany(pari_sp av, GEN* gptr[], int n){}
 /* used in addrr */
 void stackdummy(GEN z, long l) { z[0] = evaltyp(t_VECSMALL) | evallg(l); }
 
-void sorstring(long x)
+void sorstring(ulong x)
 {
 #ifdef LONG_IS_64BIT
-  printf("%08x%08x  ", (ulong)x>>32, x & MAXHALFULONG);
+  printf("%08x%08x  ", x>>32, x & MAXHALFULONG);
 #else
   printf("%08lx  ", x);
 #endif
